% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_additional_functions.R
\name{clean_pbp}
\alias{clean_pbp}
\title{Clean Play by Play Data}
\usage{
clean_pbp(pbp)
}
\arguments{
\item{pbp}{is a Data frame of play-by-play data scraped using \code{\link{fast_scraper}}.}
}
\value{
The input Data Frame of the paramter 'pbp' with the following columns
added:
\describe{
\item{ep}{expected points.}
\item{no_score_prob}{probability of no more scoring this half.}
\item{opp_fg_prob}{probability next score opponent field goal this half.}
\item{opp_safety_prob}{probability next score opponent safety  this half.}
\item{opp_td_prob}{probability of next score opponent touchdown this half.}
\item{fg_prob}{probability next score field goal this half.}
\item{safety_prob}{probability next score safety this half.}
\item{td_prob}{probability text score touchdown this half.}
}
}
\description{
Clean Play by Play Data
}
\details{
Build columns that capture what happens on all plays, including
penalties, using string extraction from play description.
Loosely based on Ben's nflscrapR guide (https://gist.github.com/guga31bb/5634562c5a2a7b1e9961ac9b6c568701)
but updated to work with the RS data, which has a different player format in
the play description; e.g. 24-M.Lynch instead of M.Lynch.
The function also standardizes team abbreviations so that, for example,
the Chargers are always represented by 'LAC' regardless of which year it was.
}
