% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_database_functions.R
\name{update_db}
\alias{update_db}
\title{Update or create a nflfastR play by play database}
\usage{
update_db(
  dbdir = ".",
  dbname = "pbp_db",
  tblname = "nflfastR_pbp",
  force_rebuild = FALSE
)
}
\arguments{
\item{dbdir}{Directory in which the database is or shall be located}

\item{dbname}{File name of an existing or desired database within dbdir}

\item{tblname}{The name of the play by play data table within the database}

\item{force_rebuild}{Logical parameter to rebuild the play by play data table
within the database from scratch in case the cleaned data were updated}
}
\description{
\code{update_db} Updates a RSQLite database with \code{nflfastR} data of all
completed games since 1999.
}
